Introduce typescript to ControlGroup.
