+++ lib/components/ControlGroup/ControlGroupDefault/styles.ts
@@ -1,0 +1,10 @@
+import { Styles } from 'jss';
+
+const styles: Styles = (): Styles => ({
+  ControlGroup: {
+    display: 'flex',
+    marginBottom: '18px'
+  }
+});
+
+export default styles;
