+++ lib/components/ControlGroup/README.md
@@ -3,36 +3,36 @@ import { Button, Form, TextBox } from 'thenativeweb-ux';
 
 <Form onSubmit={event => { event.preventDefault(); console.log('submitted') }}>
   <ControlGroup>
-    <ControlGroup.Item label='First name' adjust='flex'>
+    <ControlGroupItem label='First name' adjust='flex'>
       <TextBox
         name='first-name'
         placeholder='Your first name'
         onChange={() => {}}
       />
-    </ControlGroup.Item>
+    </ControlGroupItem>
 
-    <ControlGroup.Item label='Your name' adjust='flex'>
+    <ControlGroupItem label='Your name' adjust='flex'>
       <TextBox
         name='lastn-name'
         placeholder='Your last name'
         onChange={() => {}}
       />
-    </ControlGroup.Item>
+    </ControlGroupItem>
   </ControlGroup>
 
 
   <ControlGroup>
-    <ControlGroup.Item label='Address' adjust='flex'>
+    <ControlGroupItem label='Address' adjust='flex'>
       <TextBox
         name='adress'
         value={ 'Feuerseeplatz 14' }
         placeholder='Your business adress'
         onChange={() => {}}
       />
-      </ControlGroup.Item>
+      </ControlGroupItem>
   </ControlGroup>
 
-  <ControlGroup.Divider />
+  <ControlGroupDivider />
 
   <Button>Submit</Button>
 </Form>
